Simplify channel management: remove dead ChannelController, expose Start in None state
- Remove Create/Release from ChannelAction enum (internal lifecycle, not UI actions)
- Merge None/Idle in GetAllowedActions() so Start is always available when not active
- Delete ChannelController (/api/channel/*) — dead code never called by the UI
- Add DELETE /api/acquisition to AcquisitionController to replace channel release endpoint
- Fix unit tests: update assertions that referenced removed Create/Release actions
