--- src/PeanutVision.Api/Controllers/ChannelController.cs
@@ -1,230 +1,0 @@
-using Microsoft.AspNetCore.Mvc;
-using PeanutVision.Api.Services;
-using PeanutVision.MultiCamDriver.Imaging;
-using PeanutVision.MultiCamDriver.Imaging.Encoders;
-
-namespace PeanutVision.Api.Controllers;
-
-[ApiController]
-[Route("api/channel")]
-public class ChannelController : ControllerBase
-{
-    private readonly IAcquisitionService _acquisition;
-    private readonly string? _imageOutputDirectory;
-
-    public ChannelController(IAcquisitionService acquisition, IConfiguration configuration, IWebHostEnvironment environment)
-    {
-        _acquisition = acquisition;
-
-        var outputDir = configuration["ImageOutputDirectory"];
-        if (!string.IsNullOrEmpty(outputDir))
-        {
-            _imageOutputDirectory = Path.IsPathRooted(outputDir)
-                ? outputDir
-                : Path.Combine(environment.ContentRootPath, outputDir);
-        }
-    }
-
-    [HttpPost]
-    public ActionResult CreateChannel([FromBody] CreateChannelRequest request)
-    {
-        try
-        {
-            var profileId = new ProfileId(request.ProfileId);
-            var triggerMode = request.TriggerMode is not null
-                ? TriggerMode.Parse(request.TriggerMode)
-                : (TriggerMode?)null;
-
-            _acquisition.CreateChannel(profileId, triggerMode);
-            return Ok(new { message = "Channel created", profileId = profileId.Value });
-        }
-        catch (ArgumentException ex)
-        {
-            return BadRequest(new { error = ex.Message });
-        }
-        catch (KeyNotFoundException ex)
-        {
-            return NotFound(new { error = ex.Message });
-        }
-        catch (InvalidOperationException ex)
-        {
-            return Conflict(new { error = ex.Message });
-        }
-    }
-
-    [HttpDelete]
-    public ActionResult ReleaseChannel()
-    {
-        try
-        {
-            _acquisition.ReleaseChannel();
-            return Ok(new { message = "Channel released" });
-        }
-        catch (InvalidOperationException ex)
-        {
-            return Conflict(new { error = ex.Message });
-        }
-    }
-
-    [HttpGet]
-    public ActionResult GetStatus()
-    {
-        var stats = _acquisition.GetStatistics();
-        return Ok(new
-        {
-            state = _acquisition.ChannelState.ToString().ToLowerInvariant(),
-            profileId = _acquisition.ActiveProfileId?.Value,
-            triggerMode = _acquisition.ChannelTriggerMode?.ToString(),
-            hasFrame = _acquisition.HasFrame,
-            lastError = _acquisition.LastError,
-            allowedActions = _acquisition.GetAllowedActions(),
-            statistics = stats.HasValue
-                ? new
-                {
-                    frameCount = stats.Value.FrameCount,
-                    droppedFrameCount = stats.Value.DroppedFrameCount,
-                    errorCount = stats.Value.ErrorCount,
-                    elapsedMs = stats.Value.ElapsedTime.TotalMilliseconds,
-                    averageFps = Math.Round(stats.Value.AverageFps, 2),
-                    minFrameIntervalMs = Math.Round(stats.Value.MinFrameIntervalMs, 2),
-                    maxFrameIntervalMs = Math.Round(stats.Value.MaxFrameIntervalMs, 2),
-                    averageFrameIntervalMs = Math.Round(stats.Value.AverageFrameIntervalMs, 2),
-                    copyDropCount = stats.Value.CopyDropCount,
-                    clusterUnavailableCount = stats.Value.ClusterUnavailableCount,
-                }
-                : null,
-            recentEvents = _acquisition.GetRecentEvents(50).Select(e => new
-            {
-                timestamp = e.Timestamp,
-                type = e.Type.ToString(),
-                message = e.Message,
-            }),
-        });
-    }
-
-    [HttpPost("start")]
-    public ActionResult StartAcquisition([FromBody] StartAcquisitionRequest? request)
-    {
-        try
-        {
-            _acquisition.Start(request?.FrameCount, request?.IntervalMs);
-            return Ok(new { message = "Acquisition started" });
-        }
-        catch (ArgumentException ex)
-        {
-            return BadRequest(new { error = ex.Message });
-        }
-        catch (InvalidOperationException ex)
-        {
-            return Conflict(new { error = ex.Message });
-        }
-    }
-
-    [HttpPost("stop")]
-    public ActionResult StopAcquisition()
-    {
-        if (_acquisition.ChannelState != Services.ChannelState.Active)
-            return Conflict(new { error = "Acquisition is not active." });
-
-        _acquisition.Stop();
-        return Ok(new { message = "Acquisition stopped" });
-    }
-
-    [HttpPost("trigger")]
-    public async Task<ActionResult> Trigger()
-    {
-        try
-        {
-            var image = await _acquisition.TriggerAndWaitAsync(5000);
-
-            if (_imageOutputDirectory is not null)
-            {
-                Directory.CreateDirectory(_imageOutputDirectory);
-                var fileName = $"trigger_{DateTime.Now:yyyyMMdd_HHmmss_fff}.png";
-                var filePath = Path.Combine(_imageOutputDirectory, fileName);
-                new ImageWriter().Save(image, filePath);
-                Response.Headers["X-Image-Path"] = filePath;
-            }
-
-            var encoder = new PngEncoder();
-            var stream = new MemoryStream();
-            encoder.Encode(image, stream);
-            stream.Position = 0;
-
-            return File(stream, "image/png", "trigger.png");
-        }
-        catch (InvalidOperationException ex)
-        {
-            return Conflict(new { error = ex.Message });
-        }
-        catch (TimeoutException ex)
-        {
-            return StatusCode(504, new { error = ex.Message });
-        }
-    }
-
-    [HttpGet("latest-frame")]
-    public ActionResult GetLatestFrame()
-    {
-        var frame = _acquisition.GetLatestFrame();
-        if (frame is null)
-            return NoContent();
-
-        var encoder = new PngEncoder();
-        var stream = new MemoryStream();
-        encoder.Encode(frame, stream);
-        stream.Position = 0;
-
-        return File(stream, "image/png", "latest.png");
-    }
-
-    [HttpPost("/api/snapshot")]
-    public ActionResult Snapshot([FromBody] SnapshotRequest request)
-    {
-        try
-        {
-            var profileId = new ProfileId(request.ProfileId);
-            var triggerMode = request.TriggerMode is not null
-                ? TriggerMode.Parse(request.TriggerMode)
-                : (TriggerMode?)null;
-
-            var image = _acquisition.Snapshot(profileId, triggerMode);
-
-            if (!string.IsNullOrWhiteSpace(request.OutputPath))
-            {
-                var writer = new ImageWriter();
-                writer.Save(image, request.OutputPath);
-            }
-
-            var encoder = new PngEncoder();
-            var stream = new MemoryStream();
-            encoder.Encode(image, stream);
-            stream.Position = 0;
-
-            return File(stream, "image/png", "snapshot.png");
-        }
-        catch (ArgumentException ex)
-        {
-            return BadRequest(new { error = ex.Message });
-        }
-        catch (KeyNotFoundException ex)
-        {
-            return NotFound(new { error = ex.Message });
-        }
-        catch (InvalidOperationException ex)
-        {
-            return Conflict(new { error = ex.Message });
-        }
-        catch (TimeoutException ex)
-        {
-            return StatusCode(504, new { error = ex.Message });
-        }
-    }
-}
-
-public class CreateChannelRequest
-{
-    public required string ProfileId { get; set; }
-    public string? TriggerMode { get; set; }
-}
-
