+++ src/PeanutVision.Api/Services/ChannelAction.cs
@@ -1,3 +1,3 @@
 namespace PeanutVision.Api.Services;
 
-public enum ChannelAction { Create, Release, Start, Stop, Trigger, Snapshot }
+public enum ChannelAction { Start, Stop, Trigger, Snapshot }
